+++ tests/Integration/Command/Share/AddTest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @copyright Copyright (c) 2021 Daniel Rudolf <nextcloud.com@daniel-rudolf.de>
  *
@@ -100,10 +101,15 @@ class AddTest extends TestCase {
 			+ count($pollData['expectedInvitations']['group'] ?? [])
 			+ count($pollData['expectedInvitations']['email'] ?? []);
 
-		$expectedInvitationShareTokens = [];
+		$expectedShares = [];
+
 		foreach ($pollData['expectedInvitations'] ?? [] as $type => $shares) {
 			foreach ($shares as $userId) {
-				$expectedInvitationShareTokens[] = $this->getShareToken($pollData['pollId'], $type, $userId);
+				if (!in_array($userId, $pollData['initialShares'][$type] ?? [])) {
+					$share = $this->createShareMock($pollData['pollId'], $type, $userId);
+					$mockedShares[$type][$userId] = $share;
+					$expectedShares[] = $share;
+				}
 			}
 		}
 
@@ -117,7 +123,7 @@ class AddTest extends TestCase {
 			->expects($this->exactly($expectedShareCount))
 			->method('add')
 			->with($pollData['pollId'], $this->logicalOr(User::TYPE, Group::TYPE, Email::TYPE), $this->anything())
-			->willReturnCallback(function (int $pollId, string $type, string $userId = '') use ($pollData): Share {
+			->willReturnCallback(function (int $pollId, string $type, string $userId = '') use ($pollData, $mockedShares): Share {
 				$userIdConstraint = $this->logicalOr(...$pollData['expectedShares'][$type] ?? []);
 				$userIdConstraint->evaluate($userId);
 
@@ -125,13 +131,13 @@ class AddTest extends TestCase {
 					throw new ShareAlreadyExistsException();
 				}
 
-				return $this->createShareMock($pollId, $type, $userId);
+				return $mockedShares[$type][$userId];
 			});
 
 		$this->shareService
 			->expects($this->exactly($expectedInvitationCount))
 			->method('sendInvitation')
-			->with($this->logicalOr(...$expectedInvitationShareTokens));
+			->with($this->logicalOr(...$expectedShares));
 
 		$command = new Add(
 			$this->pollMapper,